requirements.txt
pandas
requests
selenium
openpyxl
python-dotenv